Add Google Sign-In support
This commit introduces the ability for users to sign in via their Google accounts. A new route, '/signin/google', has been added to handle Google sign-ins. This is accompanied by the necessary functions and schema updates in the user service, controller, repositories and prisma schema. It also includes the removal of the 'loggedWithGoogle' column from the 'User' table.