Improve project service and update database schema
Unused console log statements have been removed from the project service for cleaner and more efficient code. The database schema has been updated to include 'imageUrl' in the 'Project' table and a relationship constraint to delete tags linked to a project when the project is deleted has been added. The method of updating projects in the router has been changed from PUT to PATCH. The status codes for the 'getProjects' and 'getProjectsByUserId' routes have been corrected from 201 to 200.