Refactor project update schema and route
This commit refactors the projectSchema, updating the import and method calls accordingly. It further simplifies the logic in treatUpdatedProjectTags function and modifies HTTP method for the '/statusMonitor' route. The `newProjectDataSchemaInterface` has been replaced by `updateProjectDataSchemaInterface` with all their usages updated to mirror this change. For more flexibility, the fields in the update schema have been marked as optional.